src/api/requirements.txt
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.5.7
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.7.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.9
PyJWT==2.7.0
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
ruff==0.0.272
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
uvicorn==0.22.0